% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.simulation.R
\name{weather.simulation}
\alias{weather.simulation}
\title{weather simulation function}
\usage{
weather.simulation(initial)
}
\arguments{
\item{initial:}{weather condition on the initial day}
}
\value{
the number of projected sunny days in the next 10 days, as well as the projected rainfall accumulation
}
\description{
This function simulate the weather forecast in Richmond,If a day is sunny, the probability that the next day is sunny is 0.85. If a day is rainy, the probability that the next day is rainy is 0.35. If a day is rainy, the amount of rainfall accumulation in the city is governed by an Exponential(λ=2) distribution, where the value from that distribution is the rainfall in inches. If a day is sunny, there can be no rain
}
\examples{
weather.simulation("sunny")

}
\keyword{weather.simulation}
